updated ci changed import structure for clenliness
