+++ .github/actions/run-package-tests/action.yml
@@ -55,6 +55,10 @@ inputs:
     description: 'Additional pytest options'
     default: ''
     required: false
+  extra-pip-packages:
+    description: 'Space-separated pip packages to install inside the Docker container before pytest starts'
+    default: ''
+    required: false
   container-name:
     description: 'Docker container name prefix (run-id is appended automatically)'
     required: true
@@ -134,6 +138,7 @@ runs:
         quarantined-only: ${{ inputs.quarantined-only }}
         include-files: ${{ inputs.include-files }}
         volume-mount-source: ${{ github.workspace }}
+        extra-pip-packages: ${{ inputs.extra-pip-packages }}
 
     - name: Check Test Results
       if: always()
